+++ app/api/featuredItem/route.tsx
@@ -1,8 +1,9 @@
+import { ENV_KEY } from "src/types/EnvironmentType";
 import {
   getNumDigitizedItems,
   getFeaturedImage,
   imageURL,
-} from "../../utils/utils";
+} from "../../../src/utils/utils";
 import appConfig from "appConfig";
 import { NextResponse, NextRequest } from "next/server";
 export const fetchCache = "force-no-store";
@@ -27,7 +28,7 @@ export const GET = async (request: NextRequest, response: NextResponse) => {
     ),
     uuid: featuredImageData.uuid,
     title: featuredImageData.title,
-    href: `${appConfig.DC_URL[appConfig.environment]}/items/${
+    href: `${appConfig.DC_URL[appConfig.environment as ENV_KEY]}/items/${
       featuredImageData.uuid
     }`,
   };