+++ app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from "./components/pages/homePage/homePage";
-export default function Home(props: any) {
+import HomePage from "../src/components/pages/homePage/homePage";
+export default function Home() {
   return <HomePage />;
 }