+++ app/types/ExploreFurtherData.ts
@@ -1,7 +1,0 @@
-export interface ExploreFurtherData {
-  title: string;
-  url: string;
-  description: string;
-  imgAlt: string;
-  image: string;
-}